Make the structure to call other Apps when clicking item
:Release Notes:
Make the structure to call other Apps when clicking item

:Detailed Notes:
If an item is clicked, it will call proper viewer to show the item.
But currently, since viewers have not been implemented yet, only the structure is implemented.
Also if gallery app starts with launch param, appMode and folder,
it will show the item list with that appMode and folder.

:Testing Performed:
Tested on local

:QA Notes:

:Issues Addressed:
[PLAT-131219] Implement the action when clicking a file.

Change-Id: I95017e1bf64e5d008728fc6faca98aef7f1d06bb
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.mediagallery/+/304641
Tested-by: Jongson Kim <jongson.kim@lge.com>
Reviewed-by: Jongson Kim <jongson.kim@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>